chore: rename to remark-named-snippets
